api/requirements.txt
flask
psycopg2
flask-sqlalchemy
configparser
flask-marshmallow
marshmallow-sqlalchemy